Changes SublimeGit url back to http
Some people seem to be having problems installing the package if they
are not on the bleeding edge of Package Control.

Will update to use https when a new stable package control is released.
The one on the package control website still contains the bug.
(Line 1478 in Package Control.py, domain variable doesn't exist)