Updated details url to https://github.com/matthew-dumas/syntax_definition_helper
I'm not the best at writing directions, but it should be pretty straight forward once you try it. I tried to make each step as clear as possible. 

AAAPackageDev is a great package, it was just taking me way too long to figure things out because of all of the advanced features of the language files, so I made this to restrict me to just the basics so I could get the definitions done in the limited time I had.