add new PythonStautsBarShowSymbol
When edit a long python file, and the file contains more than one class, you may need to know the code under current cursor belongs to which the class, so this plug-in is displayed class name and function name in the statusbar. like this: -------->def testclass-->def testfunc

and,this plugin cando navigete backward or navigate forward like visual studio function.