Added the `Flashback` plugin
It is basically a replacement for `git log` (in a different key binding, of course) which shows the log for the current file, but as you highlight entries, the content of your view is replaced with the content of the file as stored in the respective commit. This way you can easily browse back and forth in history **one single file** and see how it developed.