Update detail url
Since I am not quite seeing much change in this tool in the future - except a few changes to make it compatible to ST2 and other OSes, I don't see a good reason to create releases for this one. Instead, I am just referring to tree/master url to use commits as releases.