Remove SideBarEnhancements until massive privacy leak plugged
It is a huge issue that every user who installs the package has the list of every installed package sent to Kite. This could be leaking internal names of products or projects inadvertently, and I am not okay with waiting to see how it is resolved. It is a travesty that such info has been collected for months already.