Move OverrideAudit to new package repository file
This adds a new organization specific package repository file to the
channel.json file and moves the existing OverrideAudit package entry
from the o.json file in this repository to the new repository file.

The OverrideAudit entry has been augmented with a specific README URL,
home page entry and some labels.

Channel Repository Tools has been used to verify that the new
repository file in the STealthy-and-haSTy organization is valid as
well as the changed made to the channel.json and o.json in the
default channel.