rename "Elm Language Support" to "Elm Syntax Highlighting"
The initial Sublime Text package for Elm has had many maintainers
over the past few years, and as a result, it had built up a bunch of
questionable code and design choices.

The plan here is to split the important functionality into two very
simple packages:

- https://github.com/evancz/elm-syntax-highlighting
- https://github.com/evancz/elm-format-on-save

The goal is to end up with (1) an extremely reliable plugin for syntax
highlighting that will be easy to install for beginners and (2) have
a secondary plugin for people who want a bit more from their workflow.