Update MarkdownEditing entry
Appears Package Control has issues with previous version checks.

ST3176 doesn't see any tags but only branches, which makes it impossible
for users on it to opt-in to alpha releases.

Hence a `st2-2.2.10` tag was added to the most recent commit which is
compatible with ST prior to ST3176.

This commit registers `st2-` prefix for all ST builds from 2000 to 3175.

Hope, ST3176 sees tags then.