Update Ruby Slim for Sublime Text 4107+
This commit re-targets "Ruby Slim" package to SublimeText organization.
The new repository is a fork of the original ruby-slim.tmbundle to
backup its history and provide backward historic data for ST2 and 3.

The main branch however contains a completely re-written syntax
definition for Sublime Text build 4107+, which is in no form compatible
with the original package.