Update MonokaiFree version constraints
In v2.0.0 I dropped support for the old tmTheme color scheme. Originally
I thought only ST4 supported the new color schemes, but it seems that
the new color scheme works fine since at least build 3211. So I've
updated the versions constraints to allow users on >=3211 to be upgraded
to the latest version.