Merge "Syntax Highlighting for Sass" into Sass
This commit merges former "Syntax Highlighting for Sass" package into "Sass"
as "Sass" is a fork and old ST2 versions are still present.

A st2-1.0.0 tag was created in "Sass" repository to keep providing ST2
compatible packages, even though that's probably not important anymore.

Intention is to cleanup packages and get rid of heavy out-dated packages.