Deprecate PHP Haml package
This commit disables PHP Haml package for ST4.

ST3 ships its own sublime-syntax based HAML syntax and it was even rewritten
for ST4 with good chances to be a way better choice than a tmLanguage syntax,
which hasn't seen any update for 10 years.

Theoretically, we could even restrict it to <3092.